+++ src/Typin/Typin/Help/DefaultHelpWriter.cs
@@ -314,7 +314,7 @@
             foreach (CommandParameterSchema parameter in command.Parameters)
             {
                 Write(' ');
-                Write(parameter.IsScalar ? $"<{parameter.Name}>" : $"<{parameter.Name}...>");
+                Write(parameter.BindableProperty.IsScalar ? $"<{parameter.Name}>" : $"<{parameter.Name}...>");
             }
 
             // Required options
@@ -324,7 +324,7 @@
                 Write(ParametersColor, !string.IsNullOrWhiteSpace(option.Name) ? $"--{option.Name}" : $"-{option.ShortName}");
 
                 Write(' ');
-                Write(option.IsScalar ? "<value>" : "<values...>");
+                Write(option.BindableProperty.IsScalar ? "<value>" : "<values...>");
             }
 
             // Options placeholder
@@ -359,7 +359,7 @@
                 }
 
                 // Valid values
-                var validValues = parameter.GetValidValues();
+                var validValues = parameter.BindableProperty.GetValidValues();
                 if (validValues.Any())
                 {
                     Write($"Valid values: {FormatValidValues(validValues)}.");
@@ -416,7 +416,7 @@
                 }
 
                 // Valid values
-                var validValues = option.GetValidValues();
+                var validValues = option.BindableProperty.GetValidValues();
                 if (validValues.Any())
                 {
                     Write("Valid values: ");
@@ -450,7 +450,9 @@
         #endregion
 
         #region Command Children
+#if !NETSTANDARD2_0
         [SuppressMessage("Style", "IDE0057:Use range operator")]
+#endif
         private void WriteCommandChildren(CommandSchema command,
                                           IEnumerable<CommandSchema> childCommands)
         {