+++ src/Typin/Typin/Internal/Input/CommandBinder.cs
@@ -27,7 +27,7 @@
 
             // Scalar parameters
             int i = 0;
-            for (; i < parameters.Count && parameters[i].IsScalar; ++i)
+            for (; i < parameters.Count && parameters[i].BindableProperty.IsScalar; ++i)
             {
                 CommandParameterSchema parameter = parameters[i];
                 CommandParameterInput scalarInput = parameterInputs[i];
@@ -39,7 +39,7 @@
             }
 
             // Non-scalar parameter (only one is allowed)
-            if (i < parameters.Count && !parameters[i].IsScalar)
+            if (i < parameters.Count && !parameters[i].BindableProperty.IsScalar)
             {
                 CommandParameterSchema nonScalarParameter = parameters[i];
 
@@ -88,9 +88,9 @@
                 // Check fallback value
                 if (!inputsProvided &&
                     option.FallbackVariableName is string v &&
-                    optionFallbackProvider.TryGetValue(v, option.Property!.PropertyType, out string? value))
+                    optionFallbackProvider.TryGetValue(v, option.BindableProperty.Property!.PropertyType, out string? value))
                 {
-                    string[] values = option.IsScalar ? new[] { value! } : value!.Split(Path.PathSeparator);
+                    string[] values = option.BindableProperty.IsScalar ? new[] { value! } : value!.Split(Path.PathSeparator);
 
                     option.BindOn(instance, values);
                     unsetRequiredOptions.Remove(option);